moved plots out of utils
