+++ sklearn_genetic/__init__.py
@@ -1,3 +1,4 @@
 from .genetic_search import GASearchCV
+from .plots import plot_fitness_evolution, plot_search_space
 
-__all__ = ['GASearchCV']
+__all__ = ['GASearchCV', 'plot_fitness_evolution', 'plot_search_space']
